[spmd] move test file and rename to testing
test_checkpoint.py is in the wrong folder, move it over and rename
it to spmd/testing to avoid confusion
